requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
auditlog3==1.0.1
autopep8==1.5.4
Babel==2.9.1
certifi==2021.5.30
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.4
django-ckeditor==6.1.0
django-colorfield==0.4.1
django-crispy-forms==1.10.0
django-datatable-view==2.1.6
django-easy-pdf==0.1.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-softdelete==0.9.2
django-utils-six==2.0
enum34==1.1.10
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
jsonfield==3.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2-binary==2.9.5
pycodestyle==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==6.0.1
reportbro-fpdf==1.7.10
reportbro-lib==1.6.0
reportlab==3.5.67
selenium==3.141.0
sentry-sdk==1.1.0
simpleeval==0.9.13
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.6
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.9
xlwt==1.3.0
python-dotenv==1.0.0
django-post-office==3.6.3
et==0.0.2
et-xmlfile==1.0.1
whitenoise==6.4.0